projects/migrate-from-aws-to-google-cloud/s3-to-cloud-storage/python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   httpx
boto3==1.40.65
    # via -r requirements.in
botocore==1.40.65
    # via
    #   boto3
    #   s3transfer
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.186.0
    # via google-generativeai
google-auth==2.42.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-genai
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-genai==1.48.0
    # via -r requirements.in
google-generativeai==0.8.5
    # via -r requirements.in
googleapis-common-protos==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
numpy==2.3.4
    # via pandas
pandas==2.3.3
    # via -r requirements.in
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.3
    # via
    #   google-genai
    #   google-generativeai
pydantic-core==2.41.4
    # via pydantic
pygments==2.19.2
    # via rich
pyparsing==3.2.5
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
pytz==2025.2
    # via pandas
requests==2.32.5
    # via
    #   google-api-core
    #   google-genai
rich==14.2.0
    # via -r requirements.in
rsa==4.9.1
    # via google-auth
s3transfer==0.14.0
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
tenacity==9.1.2
    # via google-genai
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   google-genai
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via
    #   botocore
    #   requests
websockets==15.0.1
    # via google-genai
